Capitalized enum value names, renamed TimerStat to TimeMeasurement, had string conversion not restrict min digit count
For a better look in the inspector, I felt it better to capitalize the enum value names.

So it's harder to confuse with TimerState, I felt that TimerStat needed a solid rename.

To allow var usage, I changed minDigitCount to an IntegerData without any Range attribute. Said attribute was pretty unnecessary, given how rare it would be for a user to go with a negative value